+++ packages/fast-components/src/progress/index.ts
@@ -1,0 +1,3 @@
+export * from "./base-progress";
+export * from "./progress";
+export * from "./progress-ring";