+++ packages/fast-element/src/attributes.ts
@@ -1,11 +1,11 @@
 import { Observable } from "./observation/observable";
 
 export interface ValueConverter {
-    toView(value: any): string;
+    toView(value: any): string | null;
     fromView(value: string): any;
 }
 
-export type AttributeMode = "reflect" | "boolean" | "none";
+export type AttributeMode = "reflect" | "boolean" | "fromView";
 
 export type AttributeConfiguration = {
     property: string;
@@ -33,6 +33,24 @@ export const booleanConverter: ValueConverter = {
         }
 
         return true;
+    },
+};
+
+export const nullableNumberConverter: ValueConverter = {
+    toView(value: any): string | null {
+        if (value === null || value === undefined) {
+            return null;
+        }
+        let number = value * 1;
+        return isNaN(number) ? null : number.toString();
+    },
+
+    fromView(value: any): any {
+        if (value === null || value === undefined) {
+            return null;
+        }
+        let number = value * 1;
+        return isNaN(number) ? null : number;
     },
 };
 
@@ -101,7 +119,7 @@ export class AttributeDefinition {
     ) {
         const mode = this.mode;
 
-        if (this.guards.has(object) || mode === "none") {
+        if (this.guards.has(object) || mode === "fromView") {
             return;
         }